Add dummy REST Data
Add dummy REST DATA, change dummy interface implementation

+++ JobTracker/src/net/greenbeansit/jobtracker/shared/EmployeeService.java
@@ -57,7 +57,7 @@ public interface EmployeeService extends DirectRestService {
 	
 	@GET
 	@Path("/{employeeId}/job/{jobId}")
-	Job getJob(@PathParam("employeeId") Long employeeId, @PathParam("jobId") Long jobId);
+	Job getJob(@PathParam("employeeId") Long employeeId, @PathParam("jobId") Integer jobId);
 
 	/**
 	 * Gets all {@link ActivityReport} for the {@link Employee} with the
